moving pieces around to try and make it more consistent
