+++ src/core/index.js
@@ -1,0 +1,1 @@
+export * from './get-random-number';
