+++ src/data/scopes-data.js
@@ -1,0 +1,1 @@
+export const scopes = ['daily', 'weekly', 'monthly', 'yearly'];
