+++ src/data/tasks-data.js
@@ -1,4 +1,4 @@
-export const tasksData = {
+export const tasks = {
   daily: `
 - Read the most recent journal entry
 - Review week's schedule
@@ -8,7 +8,7 @@ export const tasksData = {
   weekly: `
 - Read the week's journal entries
 - Review the calendar for this and next week
-- Review the calendar for a monthly overview 
+- Review the calendar for a monthly overview
 - Note important takeaways from the last week
 - Note the most important items for the coming week
 - List the top 3 things that need to get done week
