--- src/stores/uiStore.js
@@ -1,10 +1,0 @@
-import { writable } from 'svelte/store';
-
-let scope = writable('daily');
-let scopes = writable(['daily', 'weekly', 'monthly', 'yearly']);
-
-function changeScope(newScope) {
-  scope.set(newScope);
-}
-
-export const uiStore = { scope, scopes, changeScope };
