--- src/utils/index.js
@@ -1,5 +1,0 @@
-export const utils = {
-  getRandomNumber(min, max) {
-    return Math.floor(Math.random() * (max - min) + min);
-  },
-};
